refactor(packages)!: spell the hls.js Mux flavor hls-js
The PRD names the engine-specific import paths `media/mux/hls-js` and `media/mux/spf`. The
flavor shipped here as `hlsjs`, so the html modules, define and CDN entries, React components,
and the tag-map flavor key all move to the hyphenated spelling.

CDN bundle names follow the filenames — the config globs each media directory rather than
listing flavors — so only React's export keys and entry glob needed editing alongside.

Left alone deliberately: `hlsjs-video`, the generic hls.js element and its published subpaths,
which is a separate naming decision on main; and `MuxDataOptions['hlsjs']`, which is
mux-embed's option key rather than ours.
