+++ packages/html/src/cdn/media/mux-audio/hls-js.ts
@@ -1,0 +1,1 @@
+import '../../../define/media/mux-audio/hls-js';
