--- packages/html/src/cdn/media/mux-audio/hlsjs.ts
@@ -1,1 +1,0 @@
-import '../../../define/media/mux-audio/hlsjs';
