+++ packages/html/src/cdn/media/mux-video/hls-js.ts
@@ -1,0 +1,1 @@
+import '../../../define/media/mux-video/hls-js';
