--- packages/html/src/cdn/media/mux-video/hlsjs.ts
@@ -1,1 +1,0 @@
-import '../../../define/media/mux-video/hlsjs';
