+++ packages/html/src/define/media/mux-video/index.ts
@@ -1,1 +1,1 @@
-export * from './hlsjs';
+export * from './hls-js';
