+++ packages/html/src/media/mux-audio/index.ts
@@ -1,1 +1,1 @@
-export * from './hlsjs-media';
+export * from './hls-js-media';
