+++ packages/html/src/media/mux-video/spf-media.ts
@@ -8,7 +8,7 @@ const MuxVideoBase = MuxVideoMixin(MediaAttachMixin(CustomMediaElement('video', 
 /**
  * `<mux-video>` over the SPF-backed Mux Media instead of the hls.js-backed one.
  *
- * Shares its name with the flavor in `./hlsjs-media` on purpose: the import path picks
+ * Shares its name with the flavor in `./hls-js-media` on purpose: the import path picks
  * the engine, and nothing else about the surface moves. Deliberately not exported
  * from this directory's barrel, so importing one flavor never pulls the other's
  * engine in with it.
