+++ packages/react/src/media/mux-audio/index.ts
@@ -1,1 +1,1 @@
-export * from './hlsjs';
+export * from './hls-js';
