+++ packages/react/tsdown.config.ts
@@ -24,7 +24,7 @@ const createConfig = (mode: PackageBuildMode): UserConfig => ({
   // Flavor modules sit beside their element's index rather than under one, so
   // they need their own entries to stay separate chunks: importing one flavor
   // must never pull the other engine in with it.
-  entry: ['src/**/index.{ts,tsx}', 'src/media/*/{hlsjs,spf}.tsx', i18nLocaleEntries],
+  entry: ['src/**/index.{ts,tsx}', 'src/media/*/{hls-js,spf}.tsx', i18nLocaleEntries],
   noExternal: [/^@videojs\/skins/],
   alias: {
     '@': new URL('./src', import.meta.url).pathname,
